feat(armor): improve v3 support and canary enhancements
- Add format write version support to canary monitor
- Improve verify command with v3 format support
- Update handlers for v3 multipart operations
- Fix import paths and add missing dependencies
- Add v3 counter and vector tests

This commit consolidates v3 format improvements before the v2 default release.

+++ internal/server/format_migration.go
@@ -220,7 +220,7 @@ func (fm *FormatMigrator) Migrate(ctx context.Context, dryRun bool, concurrency 
 			}
 
 			// Check if object version is in the include list
-			if !fm.shouldMigrateVersion(armorMeta.Version) {
+			if !fm.shouldMigrateVersion(uint8(armorMeta.Version)) {
 				result.SkippedObjects++
 				fm.advanceCursor(obj.Key)
 				continue
